requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.19
celery==3.1.17
certifi==2023.5.7
charset-normalizer==3.2.0
defusedxml==0.7.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.3
django-allauth==0.18.0
django-axes==1.3.9
django-bootstrap3==5.4.0
django-model-utils==2.2
django-pipeline==1.5.4
docutils==0.12
ephem==4.1.4
futures==2.2.0
gunicorn==19.2.1
icalendar==3.8.4
idna==3.4
kombu==3.0.37
lunardate==0.1.5
oauthlib==0.7.2
phonenumbers==8.13.16
psycopg2==2.9.6
pyCalverter==1.6.1
pyephem==9.99
python-dateutil==2.4.0
python-dotenv==0.21.1
python-openid==2.2.5
python3-openid==3.2.0
pytz==2014.10
redis==2.10.5
requests==2.31.0
requests-oauthlib==0.4.2
shortuuid==0.4.2
six==1.10.0
static3==0.5.1
stripe==5.4.0
ujson==5.7.0
-e git+https://github.com/simonhayward/ukpostcodeparser.git@c172353df4bf64d2fb8db6dfb7286ff9b921e1ec#egg=UkPostcodeParser
urllib3==1.26.15
workalendar==0.3.0
